# internal/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/healthz": {
            "get": {
                "description": "Always returns 200 OK if the service is running. Used for liveness probes.",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check (liveness)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/quotes/latest": {
            "get": {
                "description": "Returns the most recent successful quote for the given currency pair. Does NOT trigger a new fetch - only returns cached/stored data.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quotes"
                ],
                "summary": "Get latest quote for a currency pair",
                "parameters": [
                    {
                        "maxLength": 3,
                        "minLength": 3,
                        "type": "string",
                        "description": "Base currency code (3 letters)",
                        "name": "base",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maxLength": 3,
                        "minLength": 3,
                        "type": "string",
                        "description": "Quote currency code (3 letters)",
                        "name": "quote",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Latest quote found",
                        "schema": {
                            "$ref": "#/definitions/api.LatestResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid currency code format",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "No quote available for the given pair",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/quotes/update": {
            "post": {
                "description": "Initiates an asynchronous update for a currency pair. Returns immediately with an update_id for tracking. Does not block on external fetch.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quotes"
                ],
                "summary": "Request asynchronous quote update",
                "parameters": [
                    {
                        "description": "Currency pair in format XXX/YYY",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Update request accepted",
                        "schema": {
                            "$ref": "#/definitions/api.UpdateResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid currency code format",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/quotes/{update_id}": {
            "get": {
                "description": "Retrieves the status and result of a quote update request by its update_id. Returns price and timestamp when status is SUCCESS.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quotes"
                ],
                "summary": "Get quote update status and result by ID",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Update ID (UUID)",
                        "name": "update_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Quote found",
                        "schema": {
                            "$ref": "#/definitions/api.QuoteResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid update_id format",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Unknown update_id",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/readyz": {
            "get": {
                "description": "Checks connectivity to critical dependencies (Postgres and Redis). Returns 200 only when all dependencies are reachable.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Readiness check",
                "responses": {
                    "200": {
                        "description": "All dependencies ready",
                        "schema": {
                            "$ref": "#/definitions/api.ReadyResponse"
                        }
                    },
                    "503": {
                        "description": "At least one dependency unavailable",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid currency code format"
                }
            }
        },
        "api.LatestResponse": {
            "type": "object",
            "properties": {
                "base": {
                    "type": "string",
                    "example": "EUR"
                },
                "price": {
                    "type": "string",
                    "example": "18.7543"
                },
                "quote": {
                    "type": "string",
                    "example": "MXN"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2025-12-01T10:15:30Z"
                }
            }
        },
        "api.QuoteResponse": {
            "type": "object",
            "properties": {
                "base": {
                    "type": "string",
                    "example": "EUR"
                },
                "error": {
                    "type": "string",
                    "example": "Failed to fetch from provider"
                },
                "price": {
                    "type": "string",
                    "example": "18.7543"
                },
                "quote": {
                    "type": "string",
                    "example": "MXN"
                },
                "status": {
                    "type": "string",
                    "example": "SUCCESS"
                },
                "update_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2025-12-01T10:15:30Z"
                }
            }
        },
        "api.ReadyResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "ready"
                }
            }
        },
        "api.UpdateRequest": {
            "type": "object",
            "properties": {
                "pair": {
                    "type": "string",
                    "example": "EUR/MXN"
                }
            }
        },
        "api.UpdateResponse": {
            "type": "object",
            "properties": {
                "update_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                }
            }
        }
    }
}